Subversion now supports revisions on externals.
A formula using svn can now provide a spec:
    :revisions => {...revision numbers...}
that contains a mapping of revision numbers to use
for externals.

The name of the external is keyed to the revision to
use for that external.

The symbol :trunk should be used to specify the reivsion
of the main repo.

An example from the Ffmpeg formula:
    head 'svn://svn.ffmpeg.org/ffmpeg/trunk',
        :revisions => { :trunk => 22916, 'libswscale' => 31045 }

Signed-off-by: Adam Vandenberg <flangy@gmail.com>