Fix first `brew update` with Git 1.7.4+; Fixes Homebrew/homebrew#5128
Behaviour for git checkout was changed such that the update would fail because it refused to checkout files on to existing files in the working directory.

This was bad behaviour anyway, we should make efforts to keep any local modifications to the Homebrew checkout. Everything is neatly resolved if we just do a --soft reset.

Closes Homebrew/homebrew#6017.