formula.rb: Don't curl when DATA is the only patch
Executing `curl(*patches)` was failing on Leopard when `patches` was an empty
list. `git-bisect` identified 48552bb42a69ece4eb2a6d9acf81a0b47adb19c9 as the
commit that caused this bug to surface by adding `--insecure` to the arg list
on Leopard. This is odd because the following both exit with status 2:

    curl
    curl --insecure

The only apparent difference is that `curl --insecure` prints "no URL
specified!" to stderr along with the usage message.

Also renamed `patches` to `external_patches` as it is more descriptive.