superenv: More thorough scrubbing of -I & -L args
It's important that -I is scrubbed thoroughly as we add most of our important paths with -isystem and -I is handled *first*.

NOTE this indicates we should probably have used -I rather than -isystem, but too late to test all that now.

Also scrub -L for bad-paths thoroughly (evaluate realpath).