Fix broken alfred cask
The ability to link any `.app` file without specifying the full path was
removed by c244385e. This means as of version 0.18.0 of `brew-cask` the
`alfred` cask has been broken, as it raises an error due to being unable
to find `Alfred Preferences.app` (since it is not located in the root
app directory).