Added Flame
Bonjour has been described as "no-pain networking". If you've ever used
iTunes and seen shared libraries popping up without having to configure
anything, then you'll know that it really is painless. That's not to
say that it's perfect. If you see someone's sharing their music, how do
you see if they're also sharing photos with iPhoto? Just start Flame to
see all the computers on the network that are advertising services. (If
it can, it'll identify the user of each computer instead of giving you
the machine's name.) Disclose the services available from each. You can
even delve right into the service record to see the gory details, if
you like that sort of thing.