Added tox-poison
Tox is a back-end tool for secure chat and and text communication,
arising out of the post-NSA paranoia about data privacy and security.
Poison is the Mac graphical frontend to Tox. (The other apps bundled,
Silica and Hibiki, are used for developing sound schemes and visual
styles for the client, and are distributed with the package, so I
decided to include them).