Added Sequential
Sequential is an image viewer for Mac OS X. It was originally designed
for opening a folder of images and displaying them in order. It works
well for viewing comics and manga. Sequential can display folders and
archives (ZIP, RAR, CBZ and CBR) of images (including JPEG, PNG, and
GIF) and PDF files. It is able to load images on the internet from a
page or image URL. Sequential has mature full screen support and can
display images' Exif data.