Added CX
CX is a new kind of cloud experience that takes content to the next
level with a range of unique features: 10 GB free secure storage,
real-time syncing, public and private sharing, interactive group
collaboration, and intelligent discovery. Easily organize your content,
access it on the go, share it with friends or publish it online — you
decide what and when. Want to discuss a file with a team? CX does that
too with group functionality and a unique file commenting system.

Keeps files in sync across multiple devices, very similar to Dropbox,
but with much more free space, however not availaible as a portable
version as DropBox is. Manages files, photos, calendars and address
book online and offline. Start with 10GB free and get up to 16GB total
when you refer friends.