Add Launchpad Manager Yosemite
Add the Yosemite version of Launchpad Manager.

Homepage: http://launchpadmanager.com

Note that the download link on the homepage is to a PHP file rather
than pointing directly to the DMG. I figured out the direct link
through Safari's developer tools and I have tested the cask and
confirmed that it's giving me the correct, most recent version.