Correct artifact stanza in google-earth-web-plugin
It seem that the nuance of creating a plugin file in
`/Library/Internet\ Plug-ins` is handled by the maintainers with logic
they rolled into a pkg.

Despite the name of the .pkg file, it seems to have just installed the
plugin and not the full Google Earth app, so it seems that this cask is
still valid. Without this change, installation fails.