Update OmniFocus cask formatting
While submitting a homebrew versions cask for the OmniFocus beta,
I was made aware of the poor ordering of some of the stanzas. As
I had based that cask off this one, I thought it prudent to tidy
the stanzas in this cask where possible.

Note that I have left the url stanza in the wrong order so as not
to duplicate the conditional logic more than is necessary.