Add Ableton Max for Live name to cycling74-max
While the product installed is named Max, this also installs Max for Live. The
latter is likely a common search term for this product. If a full license to Max
is unavailable to the user, the product installed falls back to the
functionality of Max for Live. Additionally, on the Ableton site for Max for
Live, the same download file is listed as here with no differences.