Update RawTherapee to 5.0 (#30121)
* Bump to 5.0 for mountainlion up, bump to 4.2.1 for below

New packaging format and URL structure for the 5.0 releases
4.2.1:`http://rawtherapee.com/releases_head/mac/RawTherapee_OSX_10.7_64_4.2.1.zip`
5.0-r1:`http://www.rawtherapee.com/shared/builds/mac/RawTherapee_OSX_10.9_5.0-r1_gtk3.app.zip`

5.0 format has no `DMG` just a zipped `.app`. Also the 5.0 version is named `RawTherapee-gtk3.app`

As this is developer change, I left as is, but not sure if we need to nuke the old one as it won't over-write.
The other option is an `app 'RawTherapee-gtk3.app', target: 'RawTherapee.app'`

Either way we are left with possible duplicates of non-concurrent versions. Unless a zap (I think) is put in for the `pre-5.0` version.

* Change to >=10.8, version 5.0 only going forward

Updated to reflect focus on mountain_lion and newer support.

Formula edited, audited, style'ed, and pulled.