Upgrade Appium Desktop to v1.1.1 & github hosting (#37337)
* Upgrade Appium Desktop to v1.1.1 & github hosting

Appium has migrated their code to github. The version
in the current caskroom points to an old binary
hosted on bitbucket.  The project is now
hosted and distributed on github.

I've also changed the cask token name to avoid
confusion between 'appium', the server for mobile
automation, and 'Appium Desktop' the desktop client
that allows for access and interaction with an
appium server.

I've also deleted the outdated cask. This exact situation
wasn't covered in the CONTRIBUTING.md, but seemed like
the right way to go.

Token: appium-desktop
Version: 1.1.1

* Renaming appium-desktop to appium to conform with token naming requirements

* Renaming token for better clarity and to avoid collision/confusion with outdated  token.

* Update and rename appium-desktop.rb to appium.rb

* Update appium.rb