Update intel-haxm from 7.5.1 to 7.5.2 (#67154)
* Update intel-haxm from 7.5.1 to 7.5.2

* Update intel-haxm.rb

In the Release Notes, it mentions a prerequisite of Android SDK as copy/pasted below.
"----------------------------
Prerequisites
----------------------------
Intel(R) HAXM requires the Android SDK to be installed (version 17 or higher).
For best performance, using SDK version 20 or higher is recommended.

Please refer to the Android developer website (http://developer.android.com/sdk/) for more information."

Perhaps, this is what's failing and I changed to reflect that a dependency on 'android-sdk' is now required.

* Update intel-haxm.rb

Removing whitespace for 'brew cask style'.

* Update intel-haxm.rb