audio-hijack 4.4.0
This updates `audio-hijack` to the latest version, 4.4.0.

This also updates the `livecheck` block to check the Support Page
instead of the Sparkle feed. The Sparkle feed only returns the latest
version when an appropriate `system` value is provided but we're
using an old version that 4.4.0 doesn't support. Omitting the `system`
value returns 4.0.4 instead of 4.4.0, so we can't do that. If we use
a hardcoded `1441` value (for macOS 14.4.1), we will eventually end
up in a situation again where livecheck will always return an older
version until we update the `system` value, so that isn't a
sustainable approach.

Our options seem to be to check either the Support Center
(https://rogueamoeba.com/support/knowledgebase/?showCategory=Audio+Hijack)
or the Release Notes
(https://rogueamoeba.com/audiohijack/releasenotes.php) page. The
Release Notes are more complete but the download size will grow with
each new release, so I've opted for the Support Center page. This
isn't ideal but it works and isn't tied to a specific macOS version.