skype: update livecheck
The existing `livecheck` block URL for `skype` should redirect to
the dmg file but it currently returns an `Unable to get versions`
error. For whatever reason, the server returns a 404 (Not Found)
response unless the request uses a `GET` request. When doing that,
the URL redirects to https://get.skype.com/go/getskype-skypeformac,
which redirects to the dmg. This updates the `livecheck` block to use
the intermediate get.skype.com URL, which works like we expect.