qgroundcontrol: use versioned GitHub URL
The existing `livecheck` block for `qgroundcontrol` checks the latest
GitHub release, as the first-party website only links to an
unversioned dmg file on Cloudfront (which the cask uses). This updates
the cask to use the dmg file from the latest GitHub release, aligning
the source of the cask `url` and `livecheck` block. Besides that, the
GitHub release asset URL is versioned, so we can specify a `sha256`
instead of using `:no_check`.