get-api 1.0.9
The existing `getapi_latest` URL redirects to a versioned dmg when
retrieved using a `GET` request, so this updates the cask `url` to
use the versioned URL instead.

A recent brew change in `Utils::Curl` allows us to use `HeaderMatch`
with the `getapi_latest` URL to successfully retrieve the latest
version without checking a different source (e.g., GitHub).