windterm: update url format, livecheck
Upstream doesn't provide a macOS file with every release, so we have
to check multiple GitHub releases instead of only the "latest" one.
They also seem to regularly add patch versions to existing releases
(e.g., adding 2.6.1 files to the 2.6.0 release), so we optionally
append the tag version if it differs from the file version.