falcon-sql-client: deprecate
The GitHub repository for `falcon-sql-client` was archived on
2024-06-04. The most recent release (v4.1.0) was on 2018-12-21 and
the most recent commit was on 2020-04-16. The `README` wasn't updated
to explain the project status before the repository was archived but
it does contain an older note explaining that the company is focusing
on their Dash project instead.

Since Falcon hasn't been updated in years and the repository is now
archived, it seems like the project isn't being developed further, so
this deprecates the cask accordingly.