cloudflare-warp: update livecheck
The existing `livecheck` block for `cloudflare-warp` is currently
giving an `Unable to get versions` error. I assume that the regex
was originally intended to match a `location` header URL but the
server now gives the filename in a `content-disposition` header and
there's no leading forward slash in that context.

This addresses the issue by updating the `livecheck` block regex to
remove the leading forward slash, which will work in either scenario.