moom 4.0.1
Besides updating the cask to the newest release (4.0.1), this also
updates the `livecheck` block to handle releases where the filename
version differs from the `shortVersion` (e.g., moom400.dmg for 4.0).
This will ensure that the cask uses a version that we can simply call
`#no_dots` on without any issue. Besides that, this will account for
omitted digits that aren't zero (unlike the previous `ljust(3, "0")`
approach).

One thing to note is that this only handles the situation where the
filename has more digits than the `shortVersion`. There aren't any
existing instances where the `shortVersion` is longer than the
filename version but accounting for that is less straightforward, so
I've simply left a `TODO` for if/when it becomes necessary in the
future.