jasp: fix version, livecheck
The existing `livecheck` block for `jasp` contains a `strategy` block
that uses a conditional `version` format but this is one of the ones
where we mistakenly used `match.length`, which doesn't work as
expected. This can lead to versions with a trailing comma, which
works fine but looks strange.

In this scenario we have to check for `match[1]` for this to work how
we want. This updates the `version` and `strategy` block accordingly.