keepassxc@snapshot: update version, livecheck
All of the recent snapshot builds of `keepassxc` use a 2.8.0 version
in the filename. If we use the dmg file in the `/latest/` directory
and the version doesn't change between builds, then we will get a
checksum mismatch when the build is updated. Instead, we have to use
a fully-versioned URL that uses a build directory instead, as this
will remain static between builds.

Similarly, `/latest/` is an actual directory and not just a url that
redirects to the latest build, so checking the dmg file in that
directory won't give us the full URL with the build number in it.
With this setup, we have to find the newest build directory number
on the main snapshots page, then check the directory listing page
for that build directory to find the version in the dmg file name.