vlc@nightly: update version, livecheck
The file name version has remained consistent from the start of this
year (at least) but we should capture it in the `livecheck` regex
and include it in the cask `version`, so we can easily version bump
the cask if/when the version changes in the future.

Besides that, this reworks the `livecheck` block's `strategy` block
to better align with existing standards/patterns.