cakebrewjs: update livecheck
The existing `livecheck` block for `cakebrewjs` checks a SourceForge
RSS feed but it uses the same URL as the `Sourceforge` strategy
would use with the cask URL (other than the unnecessary trailing
question mark). This updates the `livecheck` block to allow it to use
the `Sourceforge` strategy and to bring the regex in line with
typical patterns for that strategy.