gifox: update livecheck
The existing `livecheck` block for `gifox` checks a URL that redirects
to the latest dmg file but it uses a dotless version format, so this
check naively inserts dots to create a dotted version. We avoid this
type of approach whenever possible, as the assumptions of how to split
the numbers may not hold up over time. This check won't work as
expected if any of the version parts are more than one digit, as this
regex only captures the last digit of every two digits.

This resolves the issue by checking the upstream changelog page, which
links to the dmg files and also provides versions with dots.