ripx: update livecheck
The existing `livecheck` block for `ripx` checks a page that lists
changes for each version. livecheck is giving 7.1.1 as the latest
version but this is listed as "Windows Only" (this appears to be the
first release where this has happened).

This updates the regex to only match titles that don't include
anything after text like "v7.1.0 changes (20th March 2024)". I've set
up the regex to allow for something like "macOS Only" but that's the
only exception. This addresses the issue for now but we could easily
run into unexpected issues in the future if upstream does something
different. Unfortunately, I don't see any other way around this, as
this is the only version information that upstream seems to publish.