wings3d: update livecheck
The existing `livecheck` block for `wings3d` was added in #191576 and
contains a regex but no `url`. Though this technically works, we
treat `url` as a required part of a `livecheck` block (as changes to
the checkable cask URLs can change livecheck's behavior if we don't
specify the URL). This adds `url :url` to the `livecheck` block to
make the current behavior explicit.