figma: update livecheck
The existing `livecheck` block for `figma` is returning 124.6.5
(instead of 124.7.4) as the latest version and I'm not quite sure why
that's the case now. I can at least say that the request from the app
includes additional query string string parameters and most of them
don't appear to affect the current version. However, the
`localVersion` parameter makes the endpoint return whatever version
you pass to it (i.e., `version` will be 1234 in the response if you
use `localVersion=1234`), so setting this to the cask `version` will
ensure that it won't return a lower version than what we're using.