libreoffice-still: remove cask reference
`libreoffice-still-language-pack` has a `livecheck` block that
references `libreoffice-still`, which references `libreoffice`. The
former inherits the `url` and regex` values as expected but it isn't
receiving the `strategy` block from `libreoffice-still`, so it's
effectively using the check from `libreoffice` instead of
`libreoffice-still`.

This temporarily works around the bug by replacing the `libreoffice`
reference in the `libreoffice-still` `livecheck` block with the same
`url` and `regex` values, which causes the check for
`libreoffice-still-language-pack` to return 24.8.4, as expected.