macpilot: update livecheck
The existing `livecheck` block for `macpilot` is returning an `Unable
to get versions` error because the XML response isn't a Sparkle feed
anymore. The `type=sparkle` parameter doesn't appear to make a
difference, so I've removed it here. The app's updater uses
`type=normal` but the response format is the same if the `type`
parameter is omitted.

The version is still provided in the `/root/macpilot/version` element,
so this updates the `livecheck` block to simply use the `Xml`
strategy.