mixed-in-key: update livecheck
The existing `livecheck` block for `mixed-in-key` is giving an
`Unable to get versions` error, as the URLs on the download page now
use a path like `/download/67/beta/latest?key=public`, so the related
regex in the `strategy` block isn't matching. This updates the regex
to remove the `release/` part of the URL path, as that's currently
`beta/` instead. The rest of the `livecheck` block works as expected,
so I've left it as-is for now.