grafx: update livecheck
Upstream lists a number of files for macOS on the download page
(https://pulkomandy.tk/projects/GrafX2/downloads) but it's not clear
which variant we should use. The homepage links to grafx2.eu as the
"Main website" and its download page links to a specific download for
macOS, so this updates the `livecheck` block to check that. The full
version isn't listed on the page, so we have to check the filename in
the `Content-Disposition` header of the download response.

This works for now but it's far from ideal, so we may have to revisit
this if/when it breaks or returns an inappropriate version (i.e., the
matching regex is fairly loose and may not work as expected if the
upstream text format changes).