insta360-studio 5.5.3,RC_build24
The existing `version` format for `insta360-studio` causes version
comparison issues because the hash part of the path is included as
the second part of the version. If the hash is seen as lower than the
existing version hash, then the cask version will be seen as newer
than upstream even if the upstream version is actually higher.

This updates the cask to the newest version and moves the hash part
of the version to the end, so we compare the meaningful parts of the
`version` first. The second part can vary but text like
`beta2_insta360` is seen as lower than `insta360`, so we can leave
that as it is until it causes problems (at which point we will have
to move it after the third `RC_build24` part).