kate: add shimscript
Kate has a CLI which exposes some useful functionality (most notably
the `-b` and `-i` flags). Having a shimscript present in
`$(brew --prefix)/bin/kate` makes this CLI easy to acceess.

This is implemented as a shimscript rather than a symlink because,
when a symlink is used, the CLI is unable to attach to an existing
Kate session started directly from the Application bundle.

On platforms without D-Bus, Kate uses SingleApplication, a wrapper
around QLocalServer. The QLocalServer key is the SHA-256 hash of
various pieces of app data, including the executable path (returned
from `QCoreApplication::applicationFilePath()`).

When launched from a symlink, the executable path will be different to
the raw path of the binary within the Application bundle. Thus, a
different QLocalServer key will be calculated and the two Kate
instances are unable to communicate.

This only poses issues with the CLI attaching to sessions started
through other means. The CLI can successfully talk to sessions also
started from the same symlink because the executable path matches.
For example, `kate file1.txt` and then `kate file2.txt` in a second
Terminal will open file2.txt in the same window as file1.txt, and then
exit the second instance.