astah-professional: update livecheck
The existing `livecheck` block for `astah-professional` checks the
first-party download page but it now returns an `Unable to get
versions` error. The HTML no longer contains a link to the dmg file
and upstream requires visitors to fill out a form before being
redirected to the file. This updates the `livecheck` block to check
the related URL that redirects to the latest file for macOS.