uvtools 5.1.6
`uvtools` is autobumped but the autobump workflow is failing to update
to version 5.1.6 because the app is unsigned. This updates the version
and adds a `disable!` call with the future date we've been using for
this situation. This also adds a `livecheck` block to ensure that the
cask is still checked for new versions,  otherwise this will be
skipped as deprecated (since `disable!` acts as a deprecation before
reaching the disable date).