schism-tracker 20250728
`schism-tracker` is autobumped but the autobump workflow is failing to
update to version 20250728 because the app is unsigned. This updates
the version and adds a `disable!` call with the future date we've been
using for this situation. This also adds a `livecheck` block to
ensure that the cask will continue to be updated and to avoid various
Git tags in the repository that we don't want to match (but are
currently matched by the default `Git` strategy regex).