chromedriver@beta 139.0.7258.66
`chromedriver@beta` is autobumped but the workflow is failing to
update to the latest version because the ARM binary is unsigned and
the related `disable!` call is gated to the Intel binary. This updates
the cask and moves the `disable!` call to the top level, so it applies
to both ARM and Intel.