finch: remove livecheck
The `finch` cask was previously deprecated as unsigned and a
`livecheck` block was added that replicates the default livecheck
behavior to allow the cask to continue to be checked despite the
implicit deprecation from the `disable!` call. The cask was later
undeprecated but the `livecheck` block wasn't removed in the process.
This removes the `livecheck` block, as it's no longer necessary.