izip 25.8
`izip` is autobumped but the workflow is failing to update to 25.8
because downloading the dmg file fails due to Cloudflare protections
but it works if we use a browser user agent. Besides that, it seems
like upstream has changed their versioning format and this has
affected the Sparkle feed, so we also need to adjust the `livecheck`
block.