klayout: fix legacy Catalina and Big Sur versions
Upstream has removed version 0.28.12 of `klayout` from the download
page and the files are no longer available, so installation will fail
on macOS Catalina or Big Sur. The newest available version that
supports those macOS versions is now 0.27.13. This is unfortunately
lower than 0.28.12 but this appears to be the only available option
other than dropping support for those older macOS versions.