dyad 0.18.0
`dyad` is autobumped but the workflow failed to update to version
0.18.0 because the `livecheck` block simply removes any leading `v`
from the `name` field in the JSON response and incorrectly returns
"0.18.0 - create your own library of prompts to reuse across apps!"
as the newest version. This updates the version and reworks the
`livecheck` block to match the version from the tag in the file URL
path instead.