opensuperwhisper: add livecheck
By default, livecheck checks the Git tags for `opensuperwhisper` and
is returning 0.0.5 as the newest version but there hasn't been a
release after eight hours. This adds a `livecheck` block that uses
the `GithubLatest` strategy, as there can be a notable gap between
when a version is tagged and a release is created.