hydrus-network: deprecate
Version 636 was the last release of `hydrus-network` with a macOS app,
so this deprecates the cask as discontinued. Unfortunately, the
`disable!` call overwrites the `deprecation_reason` and it will be
`:fails_gatekeeper_check` instead of `:discontinued`. This causes
livecheck to continue to check the cask, so I've added a `skip` call
to manually skip the cask until we address this `brew` bug.