nvidia-nsight-compute: update livecheck
The `livecheck` block for `nvidia-nsight-compute` checks a JSON file
that contains a lot of information on various Nvidia tools and it's
currently ~2.5 MB in size (~1.5 MB compressed). This is quite large
for a check, so this updates the `livecheck` block to check the
"Getting Started" page instead. This page links to the cask URL and
is ~18 KB compressed (i.e., less data to transfer and we don't have
to parse all that JSON, so it's a lighter check overall).