meru 3.10.1
`meru` is autobumped but the workflow failed to update to 3.10.1
because `brew audit` failed with an "Artifact defined :monterey as the
minimum macOS version but the cask declared a depends_on stanza with a
minimum macOS version of :big_sur" error. This updates the version and
`depends_on macos:` value accordingly.