archiver-app: update livecheck
The existing `livecheck` block for `archiver-app` is giving an
`Unable to get versions` error, as the URL now returns a 404 (Not
Found) response. This updates the `livecheck` block to use the URL
that the app is checking, with only the necessary query string
parameters to produce the Sparkle output (i.e., omitting `format=php`
and `version=43100` parameters).

For what it's worth, the app checks incrediblebee.com/updates.php and
that redirects to the API URL. I've opted to use the API URL for now,
to align with the existing `livecheck` block URL and avoid the
redirection. However, one benefit of minimizing the query string
parameters in this case is that we're only using the parameters that
exist in both the original and redirected URLs, so we can easily
update the `livecheck` block URL to use the original domain instead of
the API subdomain if needed in the future.