prolific-pl2303 2.2.11
`prolific-pl2303` is autobumepd but the workflow failed to update to
2.2.11 because livecheck is returning an `Unable to get versions`
error, as the `livecheck` block URL now returns a 404 (Not Found)
response. This updates the `livecheck` block URL to the page that
the generic /portfolio-item/pl2303/ URL redirects to and adjusts the
regex and `strategy` block to handle the current download file URL
format. There are some parts of the file name that may change in the
future (e.g., the `_-1` suffix) but this works for now.