testfully 1.179.0
The existing `livecheck` block for `testfully` is producing an
`Unable to get versions` error, as the URL doesn't seem to work when
fetched in livecheck (and I'm not quite sure why, as it doesn't seem
to be a user agent problem). This addresses the issue by updating the
`livecheck` block to check the JSON file that the app uses to check
for new versions. This is fairly straightforward, except that the
macOS file URL is duplicated in the darwin-aarch64 and darwin-x86_64
URLs and it uses a .tar.gz file instead of a .zip file (like on the
download page).

One other thing to note is that the download page references the
latest version as both 1.178.0 and 1.179.0 but build 472 is for
1.179.0, so this also updates the cask version (though this is still
the same file as before).