atv-remote 1.4.3
`atv-remote` is autobumped but the workflow failed to update to
version 1.4.3 because the related GitHub release uses a `v1.4.3` tag
format but the previous tag used a `v.1.4.2` format (upstream
unpredictably switches between one or the other) and the Intel dmg
file name now has an `-x64` suffix. This updates the cask accordingly
and adds a `depends_on macos:` value to resolve a `brew audit` error
("Artifact defined :monterey as the minimum macOS version but the cask
declared no minimum macOS version").