wireshark-app 4.6.0
`wireshark-app` is autobumped but the workflow failed to update to
version 4.6.0 because the related app is now a universal binary, so
there's only one dmg file for 4.6.x releases. The app still checks an
arch-specific appcast, so I've left that as-is.

With the 4.6.0 release, the "old stable version" is now 4.4.10 and the
development appcast doesn't include any versions for Catalina or
newer, as it only includes an item for 4.6.0. The stable appcast also
only includes an item for 4.6.0, so it doesn't provide the "old stable
release" version. With that in mind, I had to set the `livecheck`
block for the old stable version to check the homepage, which links to
the dmg file.

Lastly, this updates the URLs to use www.wireshark.org, as that goes
through Cloudflare (and may use a CDN), whereas the existing
2.na.dl.wireshark.org URLs seem to be region-specific. We can always
switch back if this causes a problem but this aligns `wireshark-app`
with the `wireshark-chmodbpf` cask's `url`.