elephas 11.3013
`elephas` is autobumped but the workflow failed to update to version
11.3013 because the file name now uses an encoded space as the
delimiter between the name and version. This updates the version and
modifies the `livecheck` block regex to be able to match `%20` or the
typical delimiters. This also updates the `livecheck` block's
`strategy` block logic to only use a comma-delimited `version` if the
`shortVersionString` and file name version differ (they are currently
the same).