concept2-utility 7.18.00
`concept2-utility` is autobumped but the workflow failed to update to
version 7.18.00 because the page that livecheck is checking presents
the version as 7.18. This predictably fails when used as the dotless
version in the filename because it needs a `.00` patch to be 71800.
This updates the version and modifies `livecheck` block to append
`.00` to versions that only include a major and minor. Besides that,
this also updates the `livecheck` block URL to resolve a redirection.