codelite 18.2.0
`codelite` is autobumped but the workflow failed to update to version
18.2.0 because the `livecheck` block gives an `Unable to get versions`
error. The upstream download page previously used " - Stable" text
after the version but it's currently using " - Latest", so the
existing regex doesn't match.

This updates the version and adjusts the `livecheck` block regex
accordingly. 18.2.0 requires macOS Sequoia, so I've kept 18.1.0 as a
legacy version for Sonoma.