valentina-studio: disable
valentina-db.com URLs in the `valentina-studio` cask are now
inaccessible, as the website is behind Cloudflare and the protections
prevent brew from reaching the site regardless of user agent, IP
address, etc. This means that users cannot download the dmg file to
install it, so this makes the cask effectively unusable. This removes
the `livecheck` block and disables the cask accordingly.