wpsoffice-cn 12.1.23540
`wpsoffice-cn` is autobumped but the workflow failed to update to
versions beyond 7.5.1 because the `livecheck` block produces an
"Unable to get versions" error as the homepage no longer provides
version information in the format that the regex matches. This updates
the version and adjusts the `livecheck` block regex to match the
current format for the version information on the homepage.