xtool-studio: update livecheck
The existing `livecheck` block checks the homepage for version
information but it's producing an "Unable to get versions" error, as
the dmg URL is now found in some JavaScript code (in the HTML) rather
than an `href` attribute. This removes the leading `href=.*?` to allow
the regex to match the URL outside of that specific context. The
string in the JS escapes the forward slash characters, so we also have
to account for those to ensure they're omitted from the capture
groups. In my testing, this approach should work for the previous
setup (link in an `href` attribute) as well as the current setup (URL
in JS code).