beid-viewer: update livecheck
The existing `livecheck` block for `beid-viewer` is producing an
`Unable to get versions` error, as livecheck receives HTML that only
contains some minified JavaScript. Presumably this is some sort of
intermediate verification page but, regardless, livecheck is unable to
access the actual page to be able to identify the license URL and
fetch that HTML. This addresses the issue by directly checking the
license page for this software, as the number in the URL has seemingly
remained stable for years. The previous setup may have been more
robust to changes but this works and should be fine for now (granted,
it will require a manual update if the number in the URL ever
changes).