opensesame 4.1.5
`opensesame` is autobumped but the workflow failed to update to
version 4.1.5 because the related GitHub release only provides a dmg
file for Python 3.13. This is the Python version that upstream uses
for development and testing and it seems that the dmg files for the
older Python versions are added after a delay. The cask was initially
created using version 4.0.1 and that release only provided a dmg file
for Python 3.11, so that part of that was hardcoded into the cask
`url` and it has continued using those versions even after upstream
started targeting later Python versions.

This addresses the issue by switching the `livecheck` block to check
the homepage, which links to the newest dmg file that uses the Python
version that upstream uses for development and testing. This saves us
from having to check multiple GitHub releases, which was necessary
because upstream creates GitHub releases using the stable version
format but no release assets (these are described upstream as
"unpackaged"). I've set this up to include the Python suffix in the
cask `version` (making it optional), so this approach can also ensure
that we use the dmg file for the Python version that upstream targets
(rather than getting stuck on a hardcoded Python version). This works
for now and hopefully it will continue to do so in the future but I've
added an explanatory comment before the `livecheck` block to document
the situation.