biscuit 1.2.31
`biscuit` is autobumped but the workflow failed to update to version
1.2.31 because the release tag omits the leading "v" that has been
present in previous releases. This updates the version and `url`
accordingly. This also adds an appropriate `depends_on macos:` value
to address a `brew audit` error ("Artifact defined :monterey as the
minimum macOS version but the cask declared no minimum macOS
version").